Use build tag to define a constant bool
