+++ pkg/portal/service/collaborator_noauthgearonce.go
@@ -3,11 +3,4 @@
 
 package service
 
-import (
-	"context"
-)
-
-func (s *CollaboratorService) createAccountForInvitee(ctx context.Context, actorUserID string, inviteeEmail string) (err error) {
-	// It is noop
-	return
-}
+const AUTHGEARONCE = false
